+++ CHANGELOG.md
@@ -18,6 +18,13 @@
 - The `closest()` function no longer accepts an array of elements
 - Reset fieldset min-width
 
+## WIP
+
+### Fixed
+
+- Fix Height Viewport component not observing viewport changes with `expand: true`
+- Fix passing a `vh` unit to option `scale` relates to element's height in Parallax component 
+
 ## 3.16.26 (August 28, 2023)
 
 ### Fixed